Split WorkflowActionGroupFetcher.swift into four focusedused files
Sources/RunBotCore/Runner/Services/WorkflowActionGroupFetcher.swift grew to 620 lines.
Split into logical extensions for maintainability:

Created:
- WorkflowActionGroupFetcher+Runs.swift — API DTOs, run-page decoding, dedup, coalescing
- WorkflowActionGroupFetcher+Groups.swift — group construction, sorting, titles, PR labels
- WorkflowActionGroupFetcher+Jobs.swift — per-run job fetching, cache validation, bounded refresh

Edited:
- WorkflowActionGroupFetcher.swift — retains struct, init, transport/decoder, fetch(for:cache:)

Access control: transport and decoder changed from private to internal so extension
files in separate source files can access them. No public API changes.

No functional changes — all 304 tests pass, swiftlint --strict is clean."]
